feat: BMX160 init optimization, charging LED blink, BLE name with MAC suffix, compiler warning fixes
- BMX160: explicit accel 16G/100Hz + gyro 2000dps/100Hz config, increased soft reset delay to 100ms

- StateIndicator: add P1.08 charging LED with 1s blink timer via DTS gpio-leds node

- bt_mgmt: dynamic BLE device name Xinyi_IMU_XXXX using last 2 bytes of MAC address

- Sensors (PPG/BoneConduction/Temp): fix float-to-int precision issues (1e6 -> 1000000.0 with explicit casts)

- SD Card: fix signed/unsigned comparison warnings (int -> uint32_t/size_t)

- MLX90632: fix float/double literal precision warnings

- meow_ctrl_service: fix printf format warning for float

+++ src/SD_Card/SD_Card_Manager/SD_Card_Manager.cpp
@@ -325,7 +325,7 @@ int SDCardManager::ls(char *buf, size_t *buf_size) {
           &buf[used_buf_size], remaining_buf_size, "[%s]\t%s\n",
           entry.type == FS_DIR_ENTRY_DIR ? "DIR " : "FILE", entry.name);
 
-      if (len >= remaining_buf_size) {
+      if ((size_t)len >= remaining_buf_size) {
         LOG_ERR("Failed to append to buffer, error: %d", len);
         k_mutex_unlock(&m_sem_sd_mngr_oper_ongoing);
         return -EINVAL;
